Change 'height' types from Long -> Int. Int.MaxValue is 2147483647, which is an order of magnitude more than current block heights on mainnet/testnet bitcoin (#685)

+++ chain-test/src/test/scala/org/bitcoins/chain/blockchain/ChainHandlerTest.scala
@@ -216,7 +216,7 @@ class ChainHandlerTest extends ChainUnitTest {
   final def processHeaders(
       processorF: Future[ChainHandler],
       remainingHeaders: List[BlockHeader],
-      height: Long): Future[Assertion] = {
+      height: Int): Future[Assertion] = {
     remainingHeaders match {
       case header :: headersTail =>
         val newProcessorF = processorF.flatMap(_.processHeader(header))
